models/requirements.txt
openai>=1.0.0
python-docx>=0.8.11
textract>=1.6.5
numpy>=1.21.0
PyPDF2>=3.0.0
